Move extensions to new package
