+++ library/src/main/kotlin/com/gabrielfeo/gradle/enterprise/api/extension/BuildAttributesValueExtensions.kt
@@ -1,4 +1,6 @@
-package com.gabrielfeo.gradle.enterprise.api.model
+package com.gabrielfeo.gradle.enterprise.api.extension
+
+import com.gabrielfeo.gradle.enterprise.api.model.BuildAttributesValue
 
 operator fun List<BuildAttributesValue>.get(name: String): String? {
     return find { it.name == name }?.value
